apps/runner/pyproject.toml
[project]
name = "triaige-runner"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "pydantic-settings",
  "python-dotenv",
  "qdrant-client",
  "langchain",
  "langgraph",
  "langsmith",
  "openai",
  "langchain-openai",
  "Pillow",
  "httpx",
  "asyncpg",
  "pyjwt[crypto]>=2.12.0",
  "limits>=4.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
]
